requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.9.2
python-dotenv==1.0.1

numpy==2.1.3
faiss-cpu==1.9.0.post1

sentence-transformers==3.2.1
torch>=2.2.0

httpx==0.27.2
pytest==8.3.3
ruff==0.7.4
